ai-service/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
networkx>=3.2.1
pydantic>=2.7.0
numpy>=1.26.4
httpx>=0.27.0
pymongo>=4.7.0
motor>=3.4.0
